+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## [0.0.7] - 2026-03-29
+
+- Fix [Issue 1: Data refresh in Watch View](https://github.com/gsingh704/table-view/issues/1)
+- Fix [Issue 2: Index column](https://github.com/gsingh704/table-view/issues/2)
+- Better loading and pagination
+- Added chunk size setting
+
 ## [0.0.6] - 2026-03-28
 
 - Sort and Filter